TST test nested parallel_config context manager
